% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_gen_model.R
\name{adj_gen_model}
\alias{adj_gen_model}
\title{Adjust the genetic model of one or more traits}
\usage{
adj_gen_model(genome, geno, pos.cor = TRUE)
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{geno}{Genotype data on a population to phenotype. Can be a matrix of dimensions
\code{n.ind} x \code{n.loci}, the elements of which must be z {0, 1, 2}, or a list
of such matrices.}

\item{pos.cor}{A logical indicating whether the two traits should be positively correlated.}
}
\description{
Edits the additive effect of QTL of a second or more trait by assessing the
correlation among QTL for that trait with the designated pair in the first trait.
}
\details{
Although QTL for two traits can be linked, depending on the population, they
may not be positively correlated. If a correlation between two traits is desired,
this function will edit the additive effects in all but the first trait to 
induce the desired correlation between traits.
}
