% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_functions.R
\name{calc_allele_freq}
\alias{calc_allele_freq}
\title{Calculate the allele frequencies of loci}
\usage{
calc_allele_freq(genome, pop, locus.type = c("qtl", "markers", "both"))
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{pop}{A population.}

\item{locus.type}{The type of loci for which allele frequenices should be returned. Can be \code{"qtl"},
\code{"markers"}, or \code{"both"}.}
}
\value{
A \code{list} of frequencies of the 1 allele for the loci specified.
}
\description{
Calculates the frequency of markers, QTL, or both.
}
\examples{
n.mar  <- c(505, 505, 505)
len <- c(120, 130, 140)

genome <- sim_genome(len, n.mar)

# Randomly generate 15 QTL with additive allelic effects following a
# genometric series
qtl.model <- matrix(nrow = 15, ncol = 4)
genome <- sim_gen_model(genome, qtl.model, add.dist = "geometric")

# Create a random population
pop <- sim_pop(genome = genome, n.ind = 200)

# Calculate QTL allele freqencies
calc_allele_freq(genome, pop, "qtl")
# For markers
calc_allele_freq(genome, pop, "markers")

}
