% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_crossing_block.R
\name{calc_exp_genvar}
\alias{calc_exp_genvar}
\title{Calculate the expected genetic variance in families from a crossing block}
\usage{
calc_exp_genvar(genome, pedigree, founder.pop, crossing.block)
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{pedigree}{A \code{pedigree} detailing the scheme to develop the family.
Use \code{\link{sim_pedigree}} to generate.}

\item{founder.pop}{An object of class \code{pop} with the geno information for
the parents. Additional individuals can be present in \code{parent_pop}. They
will be filtered according to the parents in the \code{crossing.block}.}

\item{crossing.block}{A crossing block detailing the crosses to make. Must be a
\code{data.frame} with 2 columns: the first gives the name of parent 1, and the 
second gives the name of parent 2. See \code{\link{sim_crossing.block}}.}
}
\description{
Calculates the expected genetic variance in a cross or crosses using a pedigree
and a crossing block.
}
