% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_functions.R
\name{calc_genoval}
\alias{calc_genoval}
\title{Hidden functions; generally not to be called by the user}
\usage{
calc_genoval(genome, geno)
}
\description{
Hidden functions; generally not to be called by the user
}
