% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_functions.R
\name{calc_haplotype_freq}
\alias{calc_haplotype_freq}
\title{Calculate the frequency of haplotypes}
\usage{
calc_haplotype_freq(genome, pop, loci.list)
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{pop}{A population.}

\item{loci.list}{A \code{list} of vectors of loci names for which haplotype frequencies should be calculated.}
}
\value{
A \code{list} of haplotype frequencies
}
\description{
Calculates the frequency of haplotypes given a list of vectors of loci.
}
\examples{
n.mar  <- c(505, 505, 505)
len <- c(120, 130, 140)

genome <- sim_genome(len, n.mar)

# Randomly generate 15 QTL with additive allelic effects following a
# genometric series
qtl.model <- matrix(nrow = 15, ncol = 4)
genome <- sim_gen_model(genome, qtl.model, add.dist = "geometric")

# Create a random population
pop <- sim_pop(genome = genome, n.ind = 200)

# Sample 2 pairs of QTL
loci.list <- replicate(n = 2, sample(qtlnames(genome), 2), simplify = FALSE)

# Calculate haplotype frequencies
calc_haplotype_freq(genome, pop, loci.list)


}
