% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_functions.R
\name{calc_LD}
\alias{calc_LD}
\title{Calculate linkage disequilibrium between loci}
\usage{
calc_LD(genome, pop, loci, measure = c("r", "D"))
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{pop}{A population.}

\item{loci}{A vector of locus names for which pairwise LD should be calculated.}

\item{measure}{The type of LD measurement. Can be \code{"r2"} for the squared Pearson
correlation coefficient or \code{"D"} for the typical calculation using the
frequency of alleles and haplotypes.}
}
\value{
A \code{matrix} of pairwise LD values.
}
\description{
Calculates linkage disequilibrium (LD; as \eqn{r^2}) between loci.
}
\examples{
n.mar  <- c(505, 505, 505)
len <- c(120, 130, 140)

genome <- sim_genome(len, n.mar)

# Randomly generate 15 QTL with additive allelic effects following a
# genometric series
qtl.model <- matrix(nrow = 15, ncol = 4)
genome <- sim_gen_model(genome, qtl.model, add.dist = "geometric")

# Create a random population
pop <- sim_pop(genome = genome, n.ind = 200)

# Calculate QTL allele freqencies
calc_LD(genome, pop, qtlnames(genome))

}
