% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_functions.R
\name{check_geno}
\alias{check_geno}
\title{Check a matrix of genotypes for completeness}
\usage{
check_geno(genome, geno)
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{geno}{Genotype data on a population to phenotype. Can be a matrix of dimensions
\code{n.ind} x \code{n.loci}, the elements of which must be z {0, 1, 2}, or a list
of such matrices.}
}
\description{
Check a matrix of genotypes for completeness
}
