% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_functions.R
\name{chrlen}
\alias{chrlen}
\title{Extract the length of chromosomes}
\usage{
chrlen(genome)
}
\arguments{
\item{genome}{An object of class \code{genome}.}
}
\value{
A vector of chromosome lengths.
}
\description{
Extract the length of chromosomes
}
