% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_functions.R
\name{chrnames}
\alias{chrnames}
\title{Extract chromosome names}
\usage{
chrnames(genome)
}
\arguments{
\item{genome}{An object of class \code{genome}.}
}
\value{
A vector of character strings (the chromosome names).
}
\description{
Extract chromosome names
}
