% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosome-class.R
\docType{class}
\name{chromosome-class}
\alias{chromosome}
\alias{chromosome-class}
\title{Class "chromosome"}
\description{
A class of one or many chromosomes found within the class \code{genome}. Each chromosome has information as to the length, number of loci, and number and effect of qtl.
}
\section{Slots}{

\describe{
\item{\code{chr.len}}{An object of class \code{numeric} giving the length of the chromosome in Morgans.}

\item{\code{n.snps}}{An object of class \code{integer} giving the number of SNP loci found on the chromosome.}

\item{\code{pos.snps}}{An object of class \code{numeric} giving the Morgan position of each SNP locus found on the chromosome.}

\item{\code{n.markers}}{An object of class \code{integer} giving the number of SNP loci that are designated as markers (i.e. not QTL).}

\item{\code{pos.markers}}{An object of class \code{numeric} giving the Morgan position of each SNP marker locus found on the chromosome.}

\item{\code{n.add.qtl}}{An object of class \code{integer} giving the number of QTL that exhibit an additive effect. Since all QTL exhibit an additive effect, this is a proxy for the number of QTL.}

\item{\code{n.dom.qtl}}{An object of class \code{integer} giving the number of QTL that exhibit some dominance effect.}

\item{\code{pos.add.qtl}}{An object of class \code{numeric} giving the Morgan position of each QTL that exhibits an additive effect. Since all QTL exhibit an additive effect, this is a proxy for the location of all QTL.}

\item{\code{pos.dom.qtl}}{An object of class \code{numeric} giving the Morgan position of each QTL that exhibits a dominance effect.}

\item{\code{qtl.effects}}{An object of class \code{list} giving the additive and domiance effects of all QTL.}
}}

