% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_qtl_geno.R
\name{fill_qtl_geno}
\alias{fill_qtl_geno}
\title{Fill in genotype data for QTL}
\usage{
fill_qtl_geno(genome, geno)
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{geno}{Genotype data on a population to phenotype. Must be a matrix of dimensions
\code{n.ind} x \code{n.marker}, the elements of which must be z {0, 1, 2}. Genotype data for
the QTL should not be included.}
}
\value{
A matrix of dimensions \code{n.ind} x \code{n.loci}, with genotypes for the QTL included.
}
\description{
Generates genotype data for the QTL in a genome.
}
\details{
For now, genotypes of the QTL are copied exactly from the genotypes of the nearest 
marker. Other options may be considered later.
}
