% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_functions.R
\name{find_markerpos}
\alias{find_markerpos}
\title{Find the position of markers in the genome}
\usage{
find_markerpos(genome, marker)
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{marker}{See \code{\link[qtl]{find.markerpos}}.}
}
\description{
Find the position of markers in the genome
}
