% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popvar_functions (1).R, R/popvar_functions.R
\name{geno_to_popvar}
\alias{geno_to_popvar}
\alias{map_to_popvar}
\alias{geno_to_popvar}
\alias{map_to_popvar}
\title{Convert data for use in PopVar}
\usage{
geno_to_popvar(genome, geno)

map_to_popvar(genome)

geno_to_popvar(genome, geno)

map_to_popvar(genome)
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{geno}{Genotype data on a population to phenotype. Must be a matrix of dimensions
\code{n.ind} x \code{n.loci}, the elements of which must be z {0, 1, 2}.}

\item{genome}{An object of class \code{genome}.}

\item{geno}{Genotype data on a population to phenotype. Must be a matrix of dimensions
\code{n.ind} x \code{n.loci}, the elements of which must be z {0, 1, 2}.}
}
\value{
A \code{data.frame} of marker genotypes ready for use in \code{\link[PopVar]{pop.predict}}.

A \code{data.frame} of marker genotypes ready for use in \code{\link[PopVar]{pop.predict}}.
}
\description{
Convert data for use in PopVar

Convert data for use in PopVar

Convert data for use in PopVar

Convert data for use in PopVar
}
\section{Functions}{
\itemize{
\item \code{map_to_popvar}: 

\item \code{map_to_popvar}: 
}}

