% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome-class.R
\docType{class}
\name{genome-class}
\alias{genome}
\alias{genome-class}
\title{Class "genome"}
\description{
A class of the genome. Used to store information for each chromosome
}
\section{Slots}{

\describe{
\item{\code{n.chr}}{An object of class \code{integer} giving the number of chromosomes in the genome.}

\item{\code{n.loci}}{An object of class \code{integer} giving the total number of SNP loci in the genome.}

\item{\code{genome.len}}{An object of class \code{numeric} giving the total Morgan length of the genome.}

\item{\code{n.qtl}}{An object of class \code{integer} giving the total number of QTL in the genome.}

\item{\code{chromosomes}}{An object of class \code{list} giving all the \code{chromosome} objects in the genome.}
}}

