% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_functions.R
\name{genotype}
\alias{genotype}
\title{Genotype a population}
\usage{
genotype(genome, pop, error.rate = 0)
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{pop}{An object of class \code{pop}.}

\item{error.rate}{The genotyping error rate. This argument is not yet operational.}
}
\value{
A matrix of dimensions \code{nind} x \code{nmarkers} in format z {-1, 0, 1}.
}
\description{
Genotype a population
}
