% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_functions.R
\name{indnames}
\alias{indnames}
\title{Names of individuals}
\usage{
indnames(pop)
}
\arguments{
\item{pop}{An object of class \code{pop}.}
}
\value{
Character vector of individual names.
}
\description{
Names of individuals
}
