% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/induce_dh.R
\name{induce_dh}
\alias{induce_dh}
\title{Generate doubled-hapoids from crossover data}
\usage{
induce_dh(xodat, pedigree)
}
\arguments{
\item{xodat}{Crossover data generated from a pedigree. See \code{\link[simcross]{sim_from_pedigree}}.}

\item{pedigree}{A \code{pedigree} detailing the scheme to develop the family.
Use \code{\link{sim_pedigree}} to generate.}
}
\value{
A new xodat object to generate doubled-haploids. Only information on the final
individuals in the pedigree is returned.
}
\description{
Generate doubled-hapoids from crossover data
}
