% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haploid_functions.R
\name{is_haploid}
\alias{is_haploid}
\alias{haploid_to_geno}
\title{Haploid genotype data}
\usage{
is_haploid(geno)

haploid_to_geno(geno)
}
\arguments{
\item{geno}{Population genotype data.}
}
\description{
Haploid genotype data
}
\details{
Haploid genotype data is stored as an array of dimensions \code{2} x 
\code{n.loci} x \code{n.ind}, the elements of which must be z {0, 1}.

Haploids can be checked by using the \code{\link{is_haploid}} function, and
can be converted to a standard genotype matrix by using the 
\code{\link{haploid_to_geno}} function.
}
\examples{

data("s2_cap_haploid")
is_haploid(s2_cap_haploid)


data("s2_cap_haploid")
geno <- haploid_to_geno(s2_cap_haploid)

}
