% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_crossing_block.R
\name{make.crossing.block}
\alias{make.crossing.block}
\title{Create a crossing block}
\usage{
make.crossing.block(parent1.lines, parent2.lines, n.crosses,
  method = "random", use.parents.once = FALSE)
}
\arguments{
\item{genome.object}{An object of class \code{genome} with declared positions of
SNP markers and QTL.}

\item{hap.genome1}{A numeric vector of a parent's 1st haploid genome that is 
the same length as the number of SNP loci in the genome. The elements of the 
vector should be 0 or 1.}

\item{hap.genome2}{A numeric vector of a parent's 1st haploid genome that is 
the same length as the number of SNP loci in the genome. The elements of the 
vector should be 0 or 1.}

\item{mutate}{A logical switch whether or not to mutate haploids before recombination.}

\item{mu.snp}{A numeric vector giving the probability of a mutation at any given
SNP marker per meiotic event. Mutations are assumed completely independent.}

\item{mu.qtl}{A numeric vector giving the probability of a mutation at any given
QTL per meiotic event. Mutations are assumed completely independent.}
}
\value{
A \code{numeric} vector representing the haploid genome of a gamete that
is the product of recombination between the two haploid genomes of a parent and 
following any mutation events.
}
\description{
Creates a crossing block of parent names for use in creating a family or population.
}
\details{
TBD
}
\examples{
n.chr = 3
chr.len = c(1.0, 1.1, 1.2)
chr.snps = c(15, 16, 17)
genome <- make.genome(n.chr = n.chr, chr.len = chr.len, chr.snps = chr.snps)
genome <- trait.architecture(genome.object = genome, n.qtl = 10)

hap1 <- sample(c(1,0), sum(chr.snps), replace = T)
hap2 <- sample(c(1,0), sum(chr.snps), replace = T)

recombine(genome, hap1, hap2, mutate = T, mu.snp = 7e-8, mu.qtl = 7e-8)

}

