% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_founders.R
\name{make.founders}
\alias{make.founders}
\title{Generate founder inbreds}
\usage{
make.founders(genome.object, method = "random", theta = 0.5)
}
\arguments{
\item{genome.object}{An object of class \code{genome}.}

\item{method}{Character describing how the allele states should be decided. 
If \code{"random"}, allele states are generating from a binomial distribution 
with success probability \code{theta}. If \code{"alternate"}, the first 
founder will have the "1" allele at odd-numbered loci, and the second founder 
will have the "1" allele at even-numbered loci.}

\item{theta}{Floating point probability of any one locus having the "1" 
allele in the first founder.}
}
\value{
A \code{matrix} of dimensions 2 x m where m is the number of loci.
}
\description{
Generate a set of two completely inbred haploid parent genomes 
as founders. The alleles of the first founder are the complement of those of 
the second founder. Allele states can be generated randomly or alternating.
}
\details{
Since the genotypes generated from this function reflect completely inbred
founder lines, only the haploid genome is returned for each inbred.
}

