% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_genome.R
\name{make.genome}
\alias{make.genome}
\title{Create a genome}
\usage{
make.genome(n.chr, chr.len, chr.snps, genetic.map = NULL)
}
\arguments{
\item{n.chr}{The integer number of chromosomes in the genome.}

\item{chr.len}{Vector specifying the Morgan length of each chromosome. Vector length must equal \code{chr.num}.}

\item{chr.snps}{Vector specifying the number of SNP loci per chromosome. Vector length must equal \code{chr.num}.}

\item{genetic.map}{List of Morgan positions of loci across the genome. List length must equal \code{chr.num} and list elements must be objects of class \code{numeric} with lengths equal to the elements in \code{chr.len}. If \code{NULL}, positions are sampled from a uniform distribution.}
}
\value{
Object of class "genome" with the number of chromosomes, the number of snps,
the Morgan length of the whole genome, the number of QTL, and a list of \code{chromosome} objects.
}
\description{
Create an object of class "genome" with initial parameters of chromosome size, number of loci on each chromsome, and the position of those loci.
}
\details{
Defining the genome is the first step in performing simulations using the \code{quantgen} package. Genomes are defined as
objects of the same name. See \code{\link[quantgen]{genome}} for information about the components of a \code{genome}
object and see \code{\link[quantgen]{chromosome}} for information about the components of a \code{chromosome} object.
}
\examples{
n.chr = 10
chr.len = seq(1.0, 1.45, by = 0.05)
chr.snps = sample(15:20, size = 10, replace = T)
genome <- make.genome(n.chr = n.chr, chr.len = chr.len, chr.snps = chr.snps)

# Using the same parameters above, but defining a genetic map manually
genetic.map <- sapply(X = 1:n.chr, FUN = function(i) runif(n = chr.snps[i], min = 0, max = chr.len[i]))
genome <- make.genome(n.chr = n.chr, chr.len = chr.len, chr.snps = chr.snps, genetic.map = genetic.map)

}

