% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pedigree.R
\name{make.pedigree}
\alias{make.pedigree}
\title{Make a pedigree structure}
\usage{
make.pedigree(n.self.gen, n.ind)
}
\arguments{
\item{n.self.gen}{The number of selfing generations. For instance, 
\code{n.self.gen = 5} would produce a F_6 family}

\item{n.ind}{The number of initial F_1 individuals that are created. By
single-seed decent, this is also the number of final inbred individuals.}
}
\description{
Creates a pedigree for a bi-parental, recombinant inbred line family.
Selfing is done in the true sense.
}

