% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_population.R
\name{make.population1}
\alias{make.population1}
\title{Make a population}
\usage{
make.population1(map, crossing.block, parent.genos, n.ind, n.self.gen, m = 0,
  p = 0, cycle.number = 1)
}
\arguments{
\item{map}{A \code{list} of genetic positions of SNPs for each chromosome.}

\item{n.ind}{The number of F_1 individuals to generate.}

\item{m}{A cross-over interference parameter. See \code{\link[simcross]{
sim_from_pedigree}}.}

\item{cycle.number}{The integer number of the breeding cycle.}

\item{parent1.genome}{A vector of the parent 1 genotypes.}

\item{parent2.genome}{See \code{parent1.genome}}

\item{n.gen}{The number of selfing generations of single-seed descent. 
For instance \code{generations = 5} would results in an F_6 population. 
For each generation, each of the F_1 individuals results in a single new 
individual that has undergone an additional generation of inbreeding.}

\item{family.number}{The integer number of the family.}
}
\description{
Takes genptype information, a crossing block, and other parameters to
create multiple inbred families into one population.
}
