% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_functions.R
\name{map_to_table}
\alias{map_to_table}
\title{Convert the genetic map in the genome to a table}
\usage{
map_to_table(genome)
}
\arguments{
\item{genome}{An object of class \code{genome}.}
}
\value{
A \code{data.frame} with two columns: chromosome name and position, with row names
equal to marker names
}
\description{
Convert the genetic map in the genome to a table
}
