% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_functions.R
\name{markernames}
\alias{markernames}
\title{Extract marker names}
\usage{
markernames(genome, chr, include.qtl = FALSE)
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{chr}{See \code{\link[qtl]{markernames}}.}

\item{include.qtl}{Logical. Should the names of QTL be returned?}
}
\value{
A vector of character strings (the marker names).
}
\description{
Extract marker names
}
