% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_functions.R
\name{nchr}
\alias{nchr}
\title{Extract the number of chromosomes in the genome}
\usage{
nchr(genome)
}
\arguments{
\item{genome}{An object of class \code{genome}.}
}
\value{
A scalar of the total number of chromosomes.
}
\description{
Extract the number of chromosomes in the genome
}
