% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_functions.R
\name{nind}
\alias{nind}
\title{Number of individuals in a population}
\usage{
nind(pop)
}
\arguments{
\item{pop}{An object of class \code{pop}.}
}
\value{
Scalar number of individuals.
}
\description{
Number of individuals in a population
}
