% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_functions.R
\name{nloci}
\alias{nloci}
\title{Extract the total number of loci in the genome}
\usage{
nloci(genome, by.chr = FALSE)
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{by.chr}{Logical. Should the number of loci per chromosome be returned?}
}
\description{
Calculate the total number of loci (markers + QTL)
}
