% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_functions.R
\name{nmar}
\alias{nmar}
\title{Extract the number of markers in the genome}
\usage{
nmar(genome, by.chr = FALSE)
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{by.chr}{Logical. Should the number of markers per chromosome be returned?}
}
\value{
If \code{by.chr = TRUE}, a vector of markers per chromosomes. If \code{by.chr = FALSE},
a scalar of the total number of markers.
}
\description{
Extract the number of markers in the genome
}
