% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_functions.R
\name{nqtl}
\alias{nqtl}
\title{Extract the number of QTL in the genome}
\usage{
nqtl(genome, by.chr = FALSE)
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{by.chr}{Logical. Should the number of QTL per chromosome be returned?}
}
\value{
If \code{by.chr = TRUE}, a vector of the QTL per chromosomes. If 
\code{by.chr = FALSE}, a scalar of the total number of QTL.
}
\description{
Extract the number of QTL in the genome
}
\details{
Only unique QTL are counted. That is, pleiotropic QTL only count once.
}
