% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_genetic_variance.R
\name{pred_genvar}
\alias{pred_genvar}
\title{Predict the genetic variance in prospective crosses}
\usage{
pred_genvar(
  genome,
  pedigree,
  training.pop,
  founder.pop,
  crossing.block,
  method = c("RRBLUP", "BRR", "BayesA", "BL", "BayesB", "BayesC"),
  n.iter = 1500,
  burn.in = 500,
  thin = 5,
  save.at = ""
)
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{pedigree}{A \code{pedigree} detailing the scheme to develop the family.
Use \code{\link{sim_pedigree}} to generate.}

\item{training.pop}{An object of class \code{pop} with the elements \code{geno} and 
\code{pheno_val}. This is used as the training population.}

\item{founder.pop}{An object of class \code{pop} with the geno information for
the parents. Additional individuals can be present in \code{parent_pop}. They
will be filtered according to the parents in the \code{crossing.block}.}

\item{crossing.block}{A crossing block detailing the crosses to make. Must be a
\code{data.frame} with 2 columns: the first gives the name of parent 1, and the 
second gives the name of parent 2. See \code{\link{sim_crossing_block}}.}

\item{method}{The statistical method to predict marker effects. If \code{"RRBLUP"}, the
\code{\link[qtl]{mixed.solve}} function is used. Otherwise, the \code{\link[BGLR]{BGLR}}
function is used.}

\item{n.iter, burn.in, thin}{Number of iterations, number of burn-ins, and thinning, respectively. See 
\code{\link[BGLR]{BGLR}}.}

\item{save.at}{See \code{\link[BGLR]{BGLR}}.}
}
\description{
Uses the expected genetic variance formula and marker effects to predict the
genetic variance and correlation in potential crosses.
}
\examples{

# Simulate a genome
n.mar  <- c(505, 505, 505)
len <- c(120, 130, 140)

genome <- sim_genome(len, n.mar)

# Simulate a quantitative trait influenced by 50 QTL
qtl.model <- matrix(NA, 50, 4)
genome <- sim_gen_model(genome = genome, qtl.model = qtl.model, 
                        add.dist = "geometric", max.qtl = 50)

# Simulate the genotypes for 8 founders
founder.pop <- sim_founders(genome = genome, n.str = 8)
training.pop <- sim_phenoval(founder.pop, h2 = 0.8)

# Generate a crossing block with 5 crosses
cb <- sim_crossing_block(parents = indnames(founder.pop), n.crosses = 5)

# Create a pedigree with 100 individuals selfed to the F_3 generation
ped <- sim_pedigree(n.par = 2, n.ind = 100, n.selfgen = 2)

pred_genvar(genome = genome, pedigree = ped, training.pop = training.pop, 
            founder.pop = founder.pop, crossing.block = cb)
                

## If two traits are present, the genetic correlation is calculated
# Simulate two quantitative traits influenced by 50 pleiotropic QTL
qtl.model <- replicate(2, matrix(NA, 50, 4), simplify = FALSE)
genome <- sim_multi_gen_model(genome = genome, qtl.model = qtl.model, corr = 0.99, 
                              prob.corr = cbind(0, 1), add.dist = "normal")

# Simulate the genotypes for 8 founders
founder.pop <- sim_founders(genome = genome, n.str = 8)
training.pop <- sim_phenoval(founder.pop, h2 = 0.8)

pred_genvar(genome = genome, pedigree = ped, training.pop = training.pop, 
            founder.pop = founder.pop, crossing.block = cb)

}
