% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_functions.R
\name{pull_genotype}
\alias{pull_genotype}
\title{Pull the genotype data for a named locus}
\usage{
pull_genotype(genome, geno, loci)
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{geno}{Genotype data on a population to phenotype. Must be a matrix of dimensions
\code{n.ind} x \code{n.loci}, the elements of which must be z {0, 1, 2}.}

\item{loci}{A character vector of loci to subset from the geno matrix.}
}
\value{
A matrix of genotypes from the geno matrix, only for the loci requested.
}
\description{
Pull the genotype data for a named locus
}
