% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_functions.R
\name{pull_perf_mar}
\alias{pull_perf_mar}
\title{Extract the perfect markers from the genome}
\usage{
pull_perf_mar(genome, by.chr = FALSE)
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{by.chr}{Logical. Should the number of perfect markers per chromosome be returned?}
}
\value{
A \code{list} of the perfect QTL.
}
\description{
Extract SNP markers that are also QTL
}
\examples{
n.mar  <- c(3, 3, 3)
len <- c(120, 130, 140)

map <- lapply(X = seq(3), FUN = function(i) 
  structure(c(20, 50, 80), names = paste("C", i, "M", seq(3), sep = "")))

genome <- sim_genome(len, n.mar, map)

# Simulate genetic architecture with one perfect marker per chromosome
chromosome <- seq(3)
pos <- rep(50, 3)
a <- rep(1, 3)
d <- 0

qtl.model <- cbind(chromosome, pos, a, d)

genome <- sim_gen_model(genome, qtl.model)
 
}
