% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_functions.R
\name{pull_plei_qtl}
\alias{pull_plei_qtl}
\title{Extract the pleiotropic QTL from the genome}
\usage{
pull_plei_qtl(genome)
}
\arguments{
\item{genome}{An object of class \code{genome}.}
}
\value{
A \code{data.frame} of the pleiotropic QTL per trait.
}
\description{
Extract the pleiotropic QTL from the genome
}
