% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_functions.R
\name{pull_qtl}
\alias{pull_qtl}
\title{Extract the QTL from a genome}
\usage{
pull_qtl(genome, unique = TRUE)
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{unique}{Logical. Should information of only the unique QTL be returned?}
}
\description{
Extract the QTL from a genome
}
