% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_functions.R
\name{qtlnames}
\alias{qtlnames}
\title{Extract the names of QTL in the genome}
\usage{
qtlnames(genome, chr)
}
\arguments{
\item{genome}{An object of class \code{genome}.}
}
\value{
If \code{by.chr = TRUE}, a vector of the QTL names.
}
\description{
Extract the names of QTL in the genome
}
\details{
Only unique QTL are counted. That is, pleiotropic QTL only count once.
}
