% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_progeny_genos.R
\name{recode_progeny_genos}
\alias{recode_progeny_genos}
\title{Recode genotypes to parental states}
\usage{
recode_progeny_genos(genome, progeny.genos, founder.genos)
}
\description{
Recodes progeny genotypes to z {0, 1, 2} format based on the genotypes
of inbred founder lines.
}
