% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombine.R, R/recombine_developmental.R
\name{recombine}
\alias{recombine}
\title{Simulate a recombined gamete from a two haploid parental genomes}
\usage{
recombine(genome.object, hap.genome1, hap.genome2, mutate = TRUE, mu.snp,
  mu.qtl)

recombine(genome.object, hap.genome1, hap.genome2, mutate = TRUE, mu.snp,
  mu.qtl)
}
\arguments{
\item{genome.object}{An object of class \code{genome} with declared positions of
SNP markers and QTL.}

\item{hap.genome1}{A numeric vector of a parent's 1st haploid genome that is 
the same length as the number of SNP loci in the genome. The elements of the 
vector should be 0 or 1.}

\item{hap.genome2}{A numeric vector of a parent's 1st haploid genome that is 
the same length as the number of SNP loci in the genome. The elements of the 
vector should be 0 or 1.}

\item{mutate}{A logical switch whether or not to mutate haploids before recombination.}

\item{mu.snp}{A numeric vector giving the probability of a mutation at any given
SNP marker per meiotic event. Mutations are assumed completely independent.}

\item{mu.qtl}{A numeric vector giving the probability of a mutation at any given
QTL per meiotic event. Mutations are assumed completely independent.}

\item{genome.object}{An object of class \code{genome} with declared positions of
SNP markers and QTL.}

\item{hap.genome1}{A numeric vector of a parent's 1st haploid genome that is 
the same length as the number of SNP loci in the genome. The elements of the 
vector should be 0 or 1.}

\item{hap.genome2}{A numeric vector of a parent's 1st haploid genome that is 
the same length as the number of SNP loci in the genome. The elements of the 
vector should be 0 or 1.}

\item{mutate}{A logical switch whether or not to mutate haploids before recombination.}

\item{mu.snp}{A numeric vector giving the probability of a mutation at any given
SNP marker per meiotic event. Mutations are assumed completely independent.}

\item{mu.qtl}{A numeric vector giving the probability of a mutation at any given
QTL per meiotic event. Mutations are assumed completely independent.}
}
\value{
A \code{numeric} vector representing the haploid genome of a gamete that
is the product of recombination between the two haploid genomes of a parent and 
following any mutation events.

A \code{numeric} vector representing the haploid genome of a gamete that
is the product of recombination between the two haploid genomes of a parent and 
following any mutation events.
}
\description{
Creates a single haploid gamete from two haploid parental genomes. Recombination is modeled on the Haldane mapping function, which assumes no cross-over interference. Number of cross-over events per chromosome is samples from a Poisson distribution with lambda = L where L is the Morgan length of the chromosome. Mutation rates on a per-SNP and per-QTL basis may be included.

Creates a single haploid gamete from two haploid parental genomes. Recombination is modeled on the Haldane mapping function, which assumes no cross-over interference. Number of cross-over events per chromosome is samples from a Poisson distribution with lambda = L where L is the Morgan length of the chromosome. Mutation rates on a per-SNP and per-QTL basis may be included.
}
\details{
TBD

TBD
}
\examples{
n.chr = 3
chr.len = c(1.0, 1.1, 1.2)
chr.snps = c(15, 16, 17)
genome <- make.genome(n.chr = n.chr, chr.len = chr.len, chr.snps = chr.snps)
genome <- trait.architecture(genome.object = genome, n.qtl = 10)

hap1 <- sample(c(1,0), sum(chr.snps), replace = T)
hap2 <- sample(c(1,0), sum(chr.snps), replace = T)

recombine(genome, hap1, hap2, mutate = T, mu.snp = 7e-8, mu.qtl = 7e-8)

n.chr = 3
chr.len = c(1.0, 1.1, 1.2)
chr.snps = c(15, 16, 17)
genome <- make.genome(n.chr = n.chr, chr.len = chr.len, chr.snps = chr.snps)
genome <- trait.architecture(genome.object = genome, n.qtl = 10)

hap1 <- sample(c(1,0), sum(chr.snps), replace = T)
hap2 <- sample(c(1,0), sum(chr.snps), replace = T)

recombine(genome, hap1, hap2, mutate = T, mu.snp = 7e-8, mu.qtl = 7e-8)

}

