% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpospois.R
\name{rpospois}
\alias{rpospois}
\title{Define a function to generate random samples from a positive-Poisson distribution}
\usage{
rpospois(n, lambda)
}
\arguments{
\item{n}{The number of random samples from the distribution}

\item{lambda}{Mean parameter}
}
\value{
A numeric vector of random samples from the distribution
}
\description{
This is a hidden utility function that generates samples based on
a positive-Poisson distribution. That is, the probability of 0 is 0.
}

