% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAP_data.R
\docType{data}
\name{s2_genos}
\alias{s2_genos}
\title{Two-Row CAP Marker Data}
\format{An object of class \code{matrix} with 182 rows and 1587 columns.}
\usage{
s2_genos
}
\description{
Two-Row CAP Marker Data
}
\keyword{datasets}

