% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAP_data.R
\docType{data}
\name{s2_snp_info}
\alias{s2_snp_info}
\title{Two-Row Barley CAP Marker Map}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1587 rows and 4 columns.}
\usage{
s2_snp_info
}
\description{
Two-Row Barley CAP Marker Map
}
\keyword{datasets}
