% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAP_data.R
\docType{data}
\name{s6_cap_genos}
\alias{s6_cap_genos}
\title{Six-Row Barley CAP Marker Data}
\format{An object of class \code{matrix} with 764 rows and 1590 columns.}
\usage{
s6_cap_genos
}
\description{
Six-Row Barley CAP Marker Data
}
\keyword{datasets}
