% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAP_data.R
\docType{data}
\name{s6_snp_info}
\alias{s6_snp_info}
\title{Six-Row Barley CAP Marker Map}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1590 rows and 4 columns.}
\usage{
s6_snp_info
}
\description{
Six-Row Barley CAP Marker Map
}
\keyword{datasets}
