% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantgen_showMethods.R
\docType{methods}
\name{show,genome-method}
\alias{show,genome-method}
\title{Determine what happens when you call on a "genome" or "chromosome" object}
\usage{
\S4method{show}{genome}(object)
}
\description{
Determine what happens when you call on a "genome" or "chromosome" object
}

