% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_family.R
\name{sim_family_cb}
\alias{sim_family_cb}
\title{Simulate multiple families from a crossing block}
\usage{
sim_family_cb(genome, pedigree, founder_pop, crossing_block, ...)
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{pedigree}{A \code{pedigree} detailing the scheme to develop the family.
Use \code{\link{sim_pedigree}} to generate.}

\item{founder_pop}{An object of class \code{pop} with the geno information for
the parents. Additional individuals can be present in \code{parent_pop}. They
will be filtered according to the parents in the \code{crossing_block}.}

\item{crossing_block}{A crossing block detailing the crosses to make. Must be a
\code{data.frame} with 2 columns: the first gives the name of parent 1, and the 
second gives the name of parent 2. See \code{\link{sim_crossing_block}}.}

\item{...}{Additional arguments passed to \code{\link{sim_family}}.}
}
\description{
Simulate multiple families from a crossing block
}
\examples{

}
