% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_founders.R
\name{sim_founders}
\alias{sim_founders}
\title{Simulate founders}
\usage{
sim_founders(genome, n.str = c("2", "4", "8"), pat.freq,
  ignore.gen.model = FALSE)
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{n.str}{The number of founders (either 2, 4, or 8)}

\item{pat.freq}{A vector of length \code{n.str}/2 + 1. Frequency of SNP genotype 
patterns in the founder, where the first vector element is the frequency of monoallelic
genotypes (i.e. fixed), and the second element is the frequency of polymorphic genotypes. 
In the case of 4 or 8 founders, the second element is the frequency of SNPs unique to 1 founder, 
and the third element is the frequency of SNPs in 2 founders. In the case of 8 founders,
the fourth element is the frequency of SNPs in 3 out of 8 founders, and the fifth element
is the frequency of SNPs in 4 out of 8 founders.}
}
\value{
See \code{\link[qtl]{simFounderSnps}}. A \code{list} of the same length as \code{map},
with each element being a matrix of 0's and 2's.
}
\description{
Simulates the SNP genotype states for a set of founders. This is a simple wrapper of 
\code{\link[qtl]{simFounderSnps}} that allows for bi-parental populations
}
\details{
See \code{\link[qtl]{simFounderSnps}}. SNPs are simulated to be linkage equilibrium.
}
\examples{
# Simulate a genome
n.mar  <- c(505, 505, 505)
len <- c(120, 130, 140)

genome <- sim_genome(len, n.mar)

# Simulate a quantitative trait influenced by 50 QTL
qtl.model <- matrix(NA, 50, 4)
genome <- sim_gen_model(genome = genome, qtl.model = qtl.model, 
                        add.dist = "geometric", max.qtl = 50)

# Simulate the founder genotypes
founder_geno <- sim_founders(genome)

}
