% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pheno.R
\name{sim_pheno}
\alias{sim_pheno}
\title{Simulate phenotypes given QTL genotypes}
\usage{
sim_pheno(genome, geno, h2, n.env = 1, n.rep = 1)
}
\arguments{
\item{genome}{A \code{genome} object.}

\item{geno}{Genotype data on a population to phenotype. Must be a matrix of dimensions
\code{n.ind} x \code{n.mar}, the elements of which must be z {0, 1, 2}.}

\item{h2}{The heritability of the trait.}

\item{n.env}{The number of environments in which to phenotype.}

\item{n.rep}{The number of replicates of each individual in each environment.}
}
\value{
A list with \code{geno_val} (the genotypic values), \code{pheno_val} (the phenotypic
observations of each rep in each environment), and \code{pheno_mean} (the mean phenotypic
value for each individual).
}
\description{
Simulates phenotypic observations given the genotypic value of individuals and the 
heritability of a quantitative trait.
}
\details{
The genotypic value of individuals is calculcated as the sum of the QTL effects
carried by each individual. The genetic variance is calculated as the variance
of these genotypic values (\eqn{V_G = var(g)}).

Environmental effects are draw from a normal distribution such that \eqn{e ~ N(0, V_E)},
where \code{V_E} is the environmental variance and is 8 times the genotypic variance.
The residual variance (\code{V_R}) is calculated from \code{h2}, where

\eqn{V_R = n.rep * n.env * (\frac{Vg}{h2} - Vg)}.
}

