% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_functions.R
\name{split_haploid}
\alias{split_haploid}
\title{Split a haploid array into chromosomes}
\usage{
split_haploid(genome, geno)
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{geno}{Array of haploid genotypes. Must be an array of dimensions \code{2} 
x \code{n.loci} x \code{n.ind}, the elements of which must be z {0, 1}.}
}
\value{
A list of haploid arrays, split by chromosome.
}
\description{
Split a haploid array into chromosomes
}
