% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_functions.R
\name{subset.pop}
\alias{subset.pop}
\title{Subset a population object for specific individuals}
\usage{
\method{subset}{pop}(pop, individual)
}
\arguments{
\item{pop}{An object of class \code{pop}.}

\item{individual}{A character of individuals to subset from the \code{pop}.}
}
\description{
Subset a population object for specific individuals
}
\details{
If \code{pheno_val} is present in the \code{pop}, the variance components are
dropped.
}
