% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_functions.R
\name{subset_pop}
\alias{subset_pop}
\title{Subset a population object for specific individuals}
\usage{
subset_pop(pop, individual)
}
\arguments{
\item{pop}{An object of class \code{pop}.}

\item{individual}{A character of individuals to subset from the \code{pop}.}
}
\description{
Subset a population object for specific individuals
}
\details{
If \code{pheno_val} is present in the \code{pop}, the variance components are
dropped.
}
\examples{

# Load some historic data
data("s2_cap_genos")
data("s2_snp_info")

# Create a genome with genetic architecture
len <- tapply(s2_snp_info$cM_pos, s2_snp_info$chrom, max)
n_mar <- tapply(s2_snp_info$cM_pos, s2_snp_info$chrom, length)
map <- lapply(split(s2_snp_info, s2_snp_info$chrom), function(chr) structure(chr$cM_pos, names = chr$rs) )

genome <- sim_genome(len = len, n.mar = n_mar, map = map)

# Simulate a a trait with 15 QTL
qtl.model <- matrix(nrow = 15, ncol = 4)

genome <- sim_gen_model(genome, qtl.model, add.dist = "geometric", max.qtl = 15)

pop <- create_pop(genome = genome, geno = s2_cap_genos)

individual <- c("2ND27380", "06WA-406.6")

}
