% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.genome}
\alias{summary.genome}
\alias{print.genome}
\title{Summarize a genome object}
\usage{
\method{summary}{genome}(x)

\method{print}{genome}(x)
}
\arguments{
\item{x}{An object of class \code{genome}.}
}
\description{
Summarize a genome object

Printing genomes
}
