% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_functions.R
\name{table_to_map}
\alias{table_to_map}
\title{Convert a table of genetic positions to a map}
\usage{
table_to_map(df)
}
\arguments{
\item{df}{A \code{data.frame} of genetic positions. The first column is the 
chromosome and the second column is the genetic position (in cM). Row names
should be the marker names.}
}
\value{
A \code{list} with length equal to the number of chromosomes. Elements in the
list are named numeric vectors of genetic positions and the names are the
marker names.
}
\description{
This function is a simple wrapper around \code{\link[qtl]{table2map}}.
}
