% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_architecture.R
\name{trait.architecture}
\alias{trait.architecture}
\title{Define the genetic architecture of a quantitative trait}
\usage{
trait.architecture(genome.object, n.qtl, qtl.ind = NULL, qtl.dom.ind = NULL,
  qtl.add.eff = "geometric", qtl.dom.eff = NULL)
}
\arguments{
\item{genome.object}{An object of class \code{genome} created by the 
\code{make.genome} function.}

\item{n.qtl}{Integer of the total number of QTL.}

\item{qtl.ind}{Numeric specifying the index of SNPs to become QTL. For 
instance, if the 1st, 25th, and 50th SNPs (which may be on different 
chromosomes) are to become QTL, then \code{qtl.ind} would take the form 
\code{c(1, 25, 50)}. If \code{NULL}, QTL are randomly sampled from all 
possible SNPs. 
Default is \code{NULL}.}

\item{qtl.dom.ind}{Numeric specifying the index of QTL exhibiting dominance. 
For instance, if 100 QTL are assigned positions by \code{qtl.pos} and the 
10th, 50th, and 70th QTL (which may be on different chromsomes) will exhibit 
dominance, then \code{qtl.dom.ind} would take the form \code{c(10, 50, 70)}. 
If \code{NULL}, no QTL will be assigned dominance effects. Default is \code{NULL}.}

\item{qtl.add.eff}{Character or list of numeric vectors specifying the additive 
allelic effects (\code{a}) of QTL. Character options include "normal" or 
"geometric." See Details for a description of these options. This argument 
may also take the form of a numeric vector of length \code{n.qtl} with the 
values of \code{a} for each QTL. See Details for more information.}

\item{qtl.dom.eff}{Numeric specifying the dominance effect (\code{d}) of QTL. 
If specified, must be of length \code{n.qtl} with QTL with no desired 
dominance effect being assigned a value of 0. If \code{NULL}, no QTL will 
exhibit dominance effects. Default is \code{NULL}.}
}
\value{
An object of class \code{genome} with the specified trait architecture attributes.
}
\description{
Take an object of class "genome" and define the genetic architecture of a quantitative trait by specifying the number of QTL, the positon of the QTL, and the allelic effects of the QTL.
}
\details{
Once a genome is created, a next step is to define the architecture of a 
simulated quantitative trait. Empirical evidence suggests that different quantitative
traits possess different genetic characteristics, including number and effect 
of quantitative trait loci (QTL). This function provides flexibility as to defining
those characteristics of a simulated trait. 

Besides providing the \code{genome} object, the only other requirement of this
function is the number of QTL. The only restriction to this input is that
it cannot be greater than the total number of SNP loci in the genome.

The remaining arguments of this function allow fine-tuning of trait architecture.
The \code{qtl.ind} arguments takes a numeric vector of indices of SNPs that will be converted to QTL. The 
remaining SNPs become markers. For instance, if the genome contains 500 SNP 
loci, and the user wants the trait to be controlled by 50 QTL, 450 SNPs will remain 
as markers. If this argument is \code{NULL}, SNPs across the genome are randomly
sampled to become QTL, regardless of the number of SNPs on each chromosome.

The \code{qtl.dom.ind} is similar to \code{qtl.ind}, but it instead defines the
index of \emph{QTL} that have a non-zero dominance (\eqn{d}) value. For instance,
if 50 QTL control your simulated trait, and you want the 10th, 50th, and 70th
QTL to have some dominance effect, the input for \code{qtl.dom.ind} would be
\code{c(10, 50, 70)}.

The genotypic values of QTL are defined by the \code{qtl.add.eff} and 
\code{qtl.dom.eff} arguments. The \code{qtl.add.eff} argument determines the
genotypic values of the two homozygotes at a particular QTL. These values are
defined relative to the midparent value (\eqn{P.bar} = 0), where the value
of the unfavorable homozygote is \eqn{-a} and the values of the favorable 
homozygote is \eqn{a}, assuming that higher trait values are favorable. The
\code{qtl.dom.eff} determines the genotypic value of the heterzygote (\eqn{d}).

Two presets are included for defining the genotypic values of QTL. The \code{"geometric"}
option sets genotypic values from a geometric series:
\deqn{a = \frac{{L - 1}{L + 1}}}{a = (L - 1) / (L + 1)}
where \eqn{L} is the number of QTL and the value of the favorable homozygote 
for the \emph{k}th QTL is \eqn{a^k} for \eqn{k = 1, 2, \ldots, L}.
The justification behind this distribution of genotypic values is many quantitative
traits seem to be controlled by few loci of major effect and many loci of 
minor effect (Lande and Thompson, 1990). Note that when choosing this option,
QTL are randomly selected to have the value \eqn{-a} or \eqn{a} for homozygotes
of the "1" allele.

The other preset is \code{"normal"}, where values are sampled from a normal
distribution such that \eqn{a_k ~ N(0,1)}.

Besides presets, a user may specify the values \eqn{a} of QTL manually.

Finally, when determining breeding values, \eqn{a} is considered the value of
homozygotes for the "1" allele at a locus.
}
\examples{
n.chr = 3
chr.len = c(1.0, 1.1, 1.2)
chr.snps = c(15, 16, 17)
genome <- make.genome(n.chr = n.chr, chr.len = chr.len, chr.snps = chr.snps)
genome <- trait.architecture(genome.object = genome, n.qtl = 10)

}

