% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_gen_model.R
\name{adj_gen_model}
\alias{adj_gen_model}
\title{Adjust the genetic model of one or more traits}
\usage{
adj_gen_model(genome, geno, prob.corr, pos.corr = TRUE)
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{geno}{Genotype data on a population to phenotype. Can be a matrix of dimensions
\code{n.ind} x \code{n.loci}, the elements of which must be z {0, 1, 2}, or a list
of such matrices.}

\item{prob.corr}{Similar to the argument in \code{\link{sim_gen_model}}, except a 
third column is required which specifies the minimum correlation (i.e. linkage
disequilibrium) between QTL of one trait and QTL of another.}

\item{pos.cor}{A logical indicating whether the two traits should be positively correlated.}
}
\description{
Edits the additive effect of QTL of a second or more trait by assessing the
correlation among QTL for that trait with the designated pair in the first trait.
}
\details{
Although QTL for two traits can be linked, depending on the population, they
may not be highly correlated. If a correlation between two traits is desired,
this function will first resample markers to become QTL based on the desired
linkaged and correlations between the genotypes at those QTL. Next, the function
will edit the additive effects in all but the first trait to 
induce the desired correlation between traits.
}
\examples{
# Load some historic data
data("s2_cap_haploid")
data("s2_snp_info")

# Create a genome
map <- lapply(split(s2_snp_info, s2_snp_info$chrom), 
              function(chr) structure(chr$cM_pos, names = chr$rs) )

genome <- sim_genome(map = map, type = "hypred")

# Simulate two traits with 15 QTL, all with pairwise linkage of <= 2 cM
qtl.model <- replicate(2, matrix(nrow = 15, ncol = 4), simplify = FALSE)
prob.corr <- cbind(2, 1)

genome <- sim_gen_model(genome, qtl.model, add.dist = "geometric", max.qtl = 15, 
prob.corr = prob.corr)

pop <- create_pop(genome = genome, geno = s2_cap_haploid)

# Genetic correlation prior to adjustment
cor(pop$geno_val$trait1, pop$geno_val$trait2)

# Adjust the genome
prob.corr <- cbind(2, 1, 0.3)
genome <- adj_gen_model(genome = genome, geno = s2_cap_haploid, prob.corr = prob.corr, 
pos.corr = T)
pop <- create_pop(genome = genome, geno = s2_cap_haploid)

# Genetic correlation after adjustment
cor(pop$geno_val$trait1, pop$geno_val$trait2)

}
