% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_functions.R
\name{calc_genoval}
\alias{calc_genoval}
\title{Calculate the genotypic value of individuals}
\usage{
calc_genoval(genome, geno)
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{geno}{Genotype data on a population to phenotype. Can be a matrix of dimensions
\code{n.ind} x \code{n.loci}, the elements of which must be z {0, 1, 2}, or a list
of such matrices.}
}
\value{
A matrix of dimensions \code{n.ind} x \code{n.trait} with genotypic values
of the individuals.
}
\description{
Calculate the genotypic value of individuals
}
\details{
Calculates the genotypic value of an individual or one or more traits by adding
the allele effects and dominance deviations of the QTL carried by that individual.
}
