% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_functions.R
\name{check_geno}
\alias{check_geno}
\title{Check a matrix of genotypes for completeness}
\usage{
check_geno(genome, geno, ignore.gen.model = FALSE)
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{geno}{Genotype data on a population to phenotype. Must be a matrix of dimensions 
\code{n.ind} x \code{n.loci}, the elements of which must be z {0, 1, 2}, or a 
list of such matrices.}

\item{ignore.gen.model}{Logical - should the genetic model be ignored when
checking the matrix of genotypes?}
}
\description{
Check a matrix of genotypes for completeness
}
\examples{

n.mar  <- c(505, 505, 505)
len <- c(120, 130, 140)

genome <- sim_genome(len, n.mar)

# Randomly generate 15 QTL with additive allelic effects following a
# genometric series
qtl.model <- matrix(nrow = 15, ncol = 4)
genome <- sim_gen_model(genome, qtl.model, add.dist = "geometric")

# Create a random population
pop <- sim_pop(genome = genome, n.ind = 200)

check_geno()

}
