% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_functions.R
\name{check_geno}
\alias{check_geno}
\title{Check a matrix of genotypes for completeness}
\usage{
check_geno(genome, geno)
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{geno}{Genotype data on a population to phenotype. If the genome type is 
\code{"pbsim"}, must be a matrix of dimensions \code{n.ind} x \code{n.loci}, 
the elements of which must be z {0, 1, 2}, or a list of such matrices. If the 
genome type is \code{"hypred"}, must be an array of dimensions \code{2} x 
\code{n.loci} x \code{n.ind}, the elements of which must be z {0, 1}.}
}
\description{
Check a matrix of genotypes for completeness
}
