% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_functions.R
\name{combine_pop}
\alias{combine_pop}
\title{Combine a list of populations}
\usage{
combine_pop(pop_list)
}
\arguments{
\item{pop_list}{A list of objects of class \code{pop}.}
}
\description{
Combine a list of populations
}
\details{
If \code{pheno_val} is present in the \code{pop}, the variance components are
dropped.
}
