% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_functions.R
\name{find_proxmarkers}
\alias{find_proxmarkers}
\title{Find the position of markers near a locus}
\usage{
find_proxmarkers(genome, marker, ...)
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{marker}{See \code{\link[qtl]{find.markerpos}}.}

\item{...}{Additional arguments. Markers within a position range of \code{marker}
can be specified using the \code{min.dist} (returns markers at \code{min.dist}
from \code{marker}) and \code{max.dist} (returns markers no more than \code{min.dist}
from \code{marker}).}
}
\description{
Finds the position of markers near a particular locus. By default, the function
returns the flanking markers, however markers within a certain range can also
be returned.
}
\examples{

# Simulate the genome
n.mar  <- c(505, 505, 505)
len <- c(120, 130, 140)

genome <- sim_genome(len, n.mar)

# Sample marker names to lookup
sample_markers <- sample(markernames(genome), size = 3)

}
