% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{pred_geno_val}
\alias{pred_geno_val}
\title{Predict genotypic values using genomewide markers}
\usage{
pred_geno_val(genome, training.pop, candidate.pop, method = c("RRBLUP",
  "BRR", "BayesA", "BL", "BayesB", "BayesC"), n.iter = 1200,
  burn.in = 200, thin = 5, save.at = "")
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{training.pop}{An object of class \code{pop} with the elements \code{geno} and 
\code{pheno_val}. This is used as the training population. If marker effects
are present, they are used.}

\item{candidate.pop}{An object of class \code{pop} with the element \code{geno}.
Genotypic values are predicted for individuals in this object.}

\item{method}{The statistical method to predict marker effects. If \code{"RRBLUP"}, the
\code{\link[qtl]{mixed.solve}} function is used. Otherwise, the \code{\link[BGLR]{BGLR}}
function is used.}

\item{n.iter, burn.in, thin}{Number of iterations, number of burn-ins, and thinning, respectively. See 
\code{\link[BGLR]{BGLR}}.}

\item{save.at}{See \code{\link[BGLR]{BGLR}}.}
}
\value{
The \code{candidate.pop} with predicted genotypic values.
}
\description{
Predict genotypic values using genomewide markers
}
\details{
The \code{training.pop} must have phenotypic values associated with each entry.
The mean phenotype is used as training data in the model. Genotypic data (excluding
QTL) are used to predict marker effects, which are then used to predict the 
genotypic value of the individuals in the \code{candidate.pop}.
}
\examples{

# Simulate a genome
n.mar  <- c(505, 505, 505)
len <- c(120, 130, 140)

genome <- sim_genome(len, n.mar)

# Simulate a quantitative trait influenced by 50 QTL
qtl.model <- matrix(NA, 50, 4)
genome <- sim_gen_model(genome = genome, qtl.model = qtl.model, 
                        add.dist = "geometric", max.qtl = 50)

# Simulate the genotypes of eight founders
founder_pop <- sim_founders(genome, n.str = 8)
founder_pop <- sim_phenoval(pop = founder_pop, h2 = 0.5)

ped <- sim_pedigree(n.ind = 100, n.selfgen = 2)

# Extract the founder names
parents <- indnames(founder_pop)

# Generate a crossing block with 5 crosses
cb <- sim_crossing_block(parents = parents, n.crosses = 5)

# Simulate the populations according to the crossing block
pop <- sim_family_cb(genome = genome, pedigree = ped, founder_pop = founder_pop, 
                     crossing_block = cb)
                     
# Use the founders as a training population for the progeny
pop <- pred_geno_val(genome = genome, training.pop = founder_pop, candidate.pop = pop)

## Alternatively, predict marker effects first, then predict genotypic values
training.pop <- pred_mar_eff(genome = genome, training.pop = founder_pop)
pop <- pred_geno_val(genome = genome, training.pop = founder_pop, candidate.pop = pop)
                     
}
