% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_crossing_block.R
\name{sim_crossing_block}
\alias{sim_crossing_block}
\title{Create a crossing block from a vector of parents}
\usage{
sim_crossing_block(parents, second.parents = NULL, n.crosses,
  scheme = c("random", "all.pairwise", "chain", "pass"),
  use.parents.once = FALSE)
}
\arguments{
\item{parents}{A \code{character} of line names to use as parents. If
\code{second.parents} is not provided, crosses are assigned from randomly sampling
the entries in \code{parents}.}

\item{second.parents}{A \code{character} of line names to use as parents. If
not \code{NULL}, must be the same length as \code{parents}. Crosses are formed
by randomly pairing the entries in \code{parents} with those in 
\code{second.parents}, unless \code{scheme = "pass"} is specified.}

\item{n.crosses}{The number of crosses to generate. Cannot be more than the
possible number of crosses.}

\item{scheme}{The mating scheme. Can be one of \code{"random"}, 
\code{"all.pairwise"}, \code{"chain"}, or \code{"pass"}. See \code{Details} 
for more information on the rules of these schemes.}

\item{use.parents.once}{\code{Logical} - should parents be used only once?}
}
\description{
Create a crossing block based on parent names, with different options for
mating scheme.
}
\details{
Several options are available to generate crossing blocks from a list of
parents. Here are the rules used for generating different crossing blocks.

If \code{second.parents = NULL}:

\itemize{
  \item{If \code{scheme = "random"} (default), crosses are randomly created
  using the parents in \code{parents}. Reciprocal crosses are excluded. If
  \code{n.crosses} is less than the total number of possible crosses
  (\eqn{total crosses = (n * (n - 1)) / 2}), then \code{n.crosses} crosses are randomly
  sampled. If \code{use.parents.once = TRUE}, then the list of crosses is
  trimmed such that parents that were already used once are not used again. This
  is ignored if the final number of crosses is less than \code{n.crosses}.}
  \item{If \code{scheme = "random"}}

}
}

