% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_family.R
\name{sim_family}
\alias{sim_family}
\title{Simulate a family from a pedigree}
\usage{
sim_family(genome, pedigree, founder_geno, ...)
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{pedigree}{An object of class \code{pedigree} detailing the scheme to 
develop the family.}

\item{founder_geno}{A matrix of dimension \code{n.founder} rows and \code{n.mar}
columns with founder genotypes. Must be coded as z = {0, 1, 2}. Can also be a list
of such matrices, where each matrix is the genotypic data from a chromosome.}

\item{...}{Additional arguments. See \code{Details}.}
}
\value{
An object of class \code{pop}.
}
\description{
Simulate a family from a pedigree
}
\details{
Other arguments can be passed to internal functions in \code{sim_family}:
\itemize{
  \item{\code{m}: The crossover interference parameter. See 
  \code{\link[simcross]{sim_from_pedigree}} for more information. }
  \item{\code{p}: The proportion of crosses from non-interference process.
  See \code{\link[simcross]{sim_from_pedigree}} for more information.}
  \item{\code{family.num}: A integer designator of the family.}
  \item{\code{cycle.num}: A integer designator of the breeding cycle.}
}
}
