% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_family.R
\name{sim_family}
\alias{sim_family}
\title{Simulate a family from a pedigree}
\usage{
sim_family(genome, pedigree, founder_geno, ...)
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{pedigree}{An object of class \code{pedigree} detailing the scheme to 
develop the family.}

\item{founder_geno}{A list of the same length as there are chromosomes in the 
\code{genome}. Each element should be a matrix with \code{n.founder} rows and \code{n.mar}
columns with founder genotypes. Must be coded as z = {0, 1, 2}.}

\item{...}{Additional arguments. See \code{Details}.}
}
\description{
Simulate a family from a pedigree
}
\details{
Other arguments can be passed to internal functions in \code{sim_family}:
\itemize{
  \item{\code{m}: The crossover interference parameter. See 
  \code{\link[simcross]{sim_from_pedigree}} for more information. }
  \item{\code{p}: The proportion of crosses from non-interference process.
  See \code{\link[simcross]{sim_from_pedigree}} for more information.}
  \item{\code{family.num}: A integer designator of the family.}
  \item{\code{cycle.num}: A integer designator of the breeding cycle.}
}
}
\examples{
# Simulate a genome
n.mar  <- c(505, 505, 505)
len <- c(120, 130, 140)

genome <- sim_genome(len, n.mar)

# Simulate a pedigree
ped <- sim_pedigree(n.ind = 50, n.bcgen = 0, n.selfgen = 2)

# Simulate the founder genotypes
founder_geno <- sim_founders(genome)

# Create the family
fam <- sim_family(genome, ped, founder_geno)
 
}
