% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_family.R
\name{sim_family_cb}
\alias{sim_family_cb}
\title{Simulate multiple families from a crossing block}
\usage{
sim_family_cb(genome, pedigree, founder_pop, crossing_block, ...)
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{pedigree}{A \code{pedigree} detailing the scheme to develop the family.
Use \code{\link{sim_pedigree}} to generate.}

\item{founder_pop}{An object of class \code{pop} with the geno information for
the parents. Additional individuals can be present in \code{parent_pop}. They
will be filtered according to the parents in the \code{crossing_block}.}

\item{crossing_block}{A crossing block detailing the crosses to make. Must be a
\code{data.frame} with 2 columns: the first gives the name of parent 1, and the 
second gives the name of parent 2. See \code{\link{sim_crossing_block}}.}

\item{...}{Additional arguments passed to \code{\link{sim_family}}.}
}
\value{
An object of class \code{pop} with the information for all individuals in 
the families specified by the crossing block.
}
\description{
Simulate multiple families from a crossing block
}
\examples{

# Simulate a genome
n.mar  <- c(505, 505, 505)
len <- c(120, 130, 140)

genome <- sim_genome(len, n.mar)

# Simulate a quantitative trait influenced by 50 QTL
qtl.model <- matrix(NA, 50, 4)
genome <- sim_gen_model(genome = genome, qtl.model = qtl.model, 
                        add.dist = "geometric", max.qtl = 50)

# Simulate the genotypes of eight founders
founder_pop <- sim_founders(genome, n.str = 8)

ped <- sim_pedigree(n.ind = 100, n.selfgen = 2)

# Extract the founder names
parents <- indnames(founder_pop)

# Generate a crossing block with 5 crosses
cb <- sim_crossing_block(parents = parents, n.crosses = 5)

# Simulate the populations according to the crossing block
pop <- sim_family_cb(genome = genome, pedigree = ped, founder_pop = founder_pop, crossing_block = cb)

}
