% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_gen_model.R
\name{sim_gen_model}
\alias{sim_gen_model}
\title{Define the genetic model of a trait}
\usage{
sim_gen_model(genome, qtl.model, ...)
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{qtl.matrix}{A matrix specifying the QTL model. Each row corresponds to
a different QTL. The first column gives the chromosome number, the second 
column gives marker index, the third column gives the additive effect of
the favorable QTL allele (\code{a}) and the fourth column gives the dominance
effect at that QTL (\code{d}). If the matrix is one of NA, QTL will be
randomly assigned based on the number of rows in the matrix. See \code{Details}
for more information.}
}
\value{
A \code{genome} object with added information for the gentic model.
}
\description{
Defines the genetic architecture of a trait.
}
\details{
To simulate QTL, genetic markers in the genome are assigned to become QTL. The 
\code{qtl.model} matrix specifies the information for this assignment. The 
first column in this matrix is the chromosome number. The second column is 
the marker index. For instance, to assign the 100th marker on a chromosome
to become a QTL, the marker index would be \code{100}. The third column is
the additive effect of the favorable QTL allele (\code{a}). Genotypes homozygous
for the favorable allele are assigned a genotypic value of \code{a} and
genotypes homozygous for the unfavorable allele are assigned a genotypic value
of \code{-a}. If dominance is absent, the genotypic value of heterozygotes is
\code{0}. The fourth column is the dominance effect at the QTL. This value can
be larger that \code{a} or smaller than \code{-a} (overdominance). The genotypic
value of heterozygotes at the QTL is \code{d}.

If the \code{qtl.model} imput is completely NA, the following rules apply:
\itemize{
  \item{QTL positions are randomly drawn, with no regard to uniformity
  over chromosomes.}
  \item{The distribution of additive effects must be provided. This can
  be included as an additional argument: \code{add.dist = "normal"} or
  \code{add.dist = "geometric"}. For a distribution of \code{"normal"},
  additive effects are generated via the \code{\link[stats]{rnorm}} function.
  For a distribution of \code{"geometric"}, additive effects are calculated for
  the k-th QTL as \eqn{a^k} where \eqn{a = (1 - L) / (1 + L)} and \eqn{L} is 
  the number of QTL (Lande and Thompson, 1990).}
  \item{Dominance effects are assumed to be zero, but can be simulated
  with the optional argument: \code{dom.dist = "normal"}. Currently, only
  normal distributions are supported.}
  }
}
\examples{
n.mar  <- c(505, 505, 505)
len <- c(120, 130, 140)

genome <- sim_genome(len, n.mar)

chromosome <- c(1, 1, 2, 2, 3, 3)
index <- c(100, 200, 150, 300, 50, 400)
a <- c(1, 0.25, 0.5, 0.25, 0.25, 0.5)
d <- 0

qtl.model <- cbind(chromosome, index, a, d)

genome <- sim_gen_model(genome, qtl.model)

}

