% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_gen_model.R
\name{sim_multi_gen_model}
\alias{sim_multi_gen_model}
\title{Define a genetic model for two or more traits}
\usage{
sim_multi_gen_model(genome, qtl.model, corr, prob.corr, ...)
}
\arguments{
\item{genome}{An object of class \code{genome}.}

\item{qtl.model}{A matrix specifying the QTL model. See \code{\link[pbsim]{sim_gen_model}}.
If the qtl.model matrices are NA, the two traits must have the same number of QTL.}

\item{corr}{The desired genetic correlation if QTL are to be drawn randomly.
May be positive or negative. See \code{Details} regarding the multivariate 
random sampling of additive effects.}

\item{prob.corr}{A matrix of two columns defining the probabilities that pairs 
of QTL for two or more traits are at most \code{x} cM apart. The first column 
sets the maximum distance between a QTL from  a second trait and a QTL from the 
first trait, and the second column is the probability that QTL from a second 
trait have that maximum distance. Pleiotropic QTL can be simulated by providing 
a 0 in the first column, and no genetic linkage is simulated if 50 is in the 
first column.}
}
\description{
Generates a genetic model for two or more traits with desired genetic correlation.
}
\details{
QTL are simulated by sampling or specifying existing markers, which become "hidden."

The \code{qtl.model} matrix specifies the information for this assignment. The 
first column in this matrix is the chromosome number. The second column is 
the QTL position. The third column is the additive effect of the "1" allele
at the QTL (\code{a}). Genotypes homozygous for the "1" allele are assigned a 
genotypic value of \code{a} and genotypes homozygous for the "-1" allele are 
assigned a genotypic value of \code{-a}. The value of \code{a} provided in 
\code{qtl.model} can be negative. The fourth column is the dominance effect at the 
QTL. If non-zero, this value can be larger that \code{a} or smaller than \code{-a} 
(overdominance). The genotypic value of heterozygotes at the QTL is \code{d}.

Other arguments include:
\describe{
  \item{\code{add.dist}}{The distribution of additive effects of QTL (if additive 
  effects are not provided in the \code{qtl.model} input). Can be 
  \code{"normal"} or \code{"geometric"}. For a distribution of \code{"normal"}, 
  additive effects are generated via a multivariate normal distribution using
  the \code{\link[mvtnorm]{rmvnorm}} function and with variance-covariance matrix
  \code{Sigma = rbind(c(1, corr), c(corr, 1))}. For a distribution of \code{"geometric"}, 
  additive effects are calculated for the k-th QTL as \eqn{a^k} where 
  \eqn{a = (1 - L) / (1 + L)} and \eqn{L} is the number of QTL (Lande and Thompson, 1990).}.
  the same variance-covariance matrix above is then used to adjust the additive
  effects to achieve the desired correlation. This approach assumes that pairs
  of QTL are in coupling phase linkage, therefore the desired correlation will
  be different than the observed correlation depending on the population.
  \item{\code{dom.dist}}{The distribution of dominance effects of QTL (if dominance
  effects are not provided in the \code{qtl.model} input). Can be 
  \code{"normal"} for normally-distributed dominance effects.}
  \item{\code{max.qtl}}{The maximum number of QTL in the simulation experiment. Must
  be passed if the QTL are randomly sampled. If a trait is controlled by \emph{L} QTL
  and \code{max.qtl = M}, then \emph{M} - \emph{L} QTL are given NULL effects.
  This is useful if you want to simulate variable genetic architecture, but keep
  the number of SNP markers constant.}
}
}
\examples{
# Simulate a genome
n.mar  <- c(505, 505, 505)
len <- c(120, 130, 140)

genome <- sim_genome(len, n.mar)

# Simulate two traits that are completely pleiotropy and influenced by 50 QTL
qtl.model <- replicate(n = 2, matrix(NA, 50, 4), simplify = FALSE)
genome <- sim_multi_gen_model(genome = genome, qtl.model = qtl.model, cor = 0.6, 
                              prob.corr = cbind(0, 1), add.dist = "geometric")


}
