% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_phenotypes.R
\name{sim_phenoval}
\alias{sim_phenoval}
\title{Simulate phenotypic values}
\usage{
sim_phenoval(pop, h2, n.env = 1, n.rep = 1, ...)
}
\arguments{
\item{pop}{An object of class \code{pop}.}

\item{h2}{The heritability of the trait or traits. May be a numeric vector of 
length 1 (heritability is the same for all traits) or a numeric vector of 
length n_trait.}

\item{n.env}{The number of environments in which to phenotype.}

\item{n.rep}{The number of replicates of each individual in each environment.}

\item{...}{Other arguments. See \code{Details}.}

\item{return.eff}{Should the variance components and effects be returned?}
}
\value{
An object of class \code{pop} with all information in the input \code{pop} object,
plus simulated phenotypes.
}
\description{
Simulates phenotypic observations given the genotypic value of individuals and the 
heritability of a quantitative trait.
}
\details{
Other arguments that can be specified are :
\describe{
  \item{\code{V_E}}{The variance of environmental effects. May be a scalar 
  (V_E is the same for all traits) or a numeric vector of length n_trait. }
  \item{\code{V_GE}}{The variance of genotype-environment interaction effects. May be a scalar 
  (V_GE is the same for all traits) or a numeric vector of length n_trait. }
  \item{\code{V_R}}{The variance of the residual effects. May be a scalar 
  (V_R is the same for all traits) or a numeric vector of length n_trait. }
}

Environmental effects are drawn from a normal distribution such that \eqn{e ~ N(0, V_E)},
where \code{V_E} is the environmental variance. If \code{V_E} is not provided, it 
is 8 times the genotypic variance. Genotype-environmental interaction effects are drawn
from a normal distribution such that \eqn{ge ~ N(0, V_GE)}, where \code{V_GE} is the genotype-environment 
interaction variance. If \code{V_GE} is not provided, it defaults to 2 times the genotypic variance. 
Residual effects are drawn from a normal distribution such that \eqn{\epsilon ~ N(0, V_R)}, where \code{V_R} is 
the residual variance. If \code{V_R} is not provided, it is calculated from \code{h2}, where 
\eqn{V_R = n.rep * n.env * (\frac{Vg}{h2} - Vg)}.
}
\examples{

# Create a genome
n.mar  <- c(505, 505, 505)
len <- c(120, 130, 140)

genome <- sim_genome(len, n.mar)

# Simulate a a trait with 15 QTL
qtl.model <- matrix(nrow = 15, ncol = 4)

genome <- sim_gen_model(genome, qtl.model, add.dist = "geometric", max.qtl = 15)

pop <- sim_pop(genome = genome, n.ind = 200)

## Default simulates environment and residual effects, but no GxE
pop1 <- sim_phenoval(pop = pop, h2 = 0.5, n.env = 2, n.rep = 2)

## Simulate

}
