% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_functions.R
\name{subset_pop}
\alias{subset_pop}
\title{Subset a population object for specific individuals}
\usage{
subset_pop(pop, individual)
}
\arguments{
\item{pop}{An object of class \code{pop}.}

\item{individual}{A character of individuals to subset from the \code{pop}.}
}
\description{
Subset a population object for specific individuals
}
\details{
If \code{pheno_val} is present in the \code{pop}, the variance components are
dropped.
}
\examples{

# Simulate a genome
n.mar  <- c(505, 505, 505)
len <- c(120, 130, 140)

genome <- sim_genome(len, n.mar)

# Simulate a quantitative trait influenced by 50 QTL
qtl.model <- matrix(NA, 50, 4)
genome <- sim_gen_model(genome = genome, qtl.model = qtl.model, 
                        add.dist = "geometric", max.qtl = 50)
                        
# Simulate the population
pop <- sim_pop(genome = genome, n.ind = 100)

# Subset the population
subset_pop(pop, c("ind001", "ind100"))

}
